src/flask/requirements.txt
dacite==1.8.0
camel-converter==3.0.0
python-environ==0.4.54
pytest-flask==1.2.0
Flask-WebTest==0.1.1
colander==2.0
flask-cors==3.0.10
Flask-HTTPAuth==4.8.0